point-view: refactor component state and lifecycle
- extract shared Network Keys logic to custom hook
- refactor API calls with useEffect hooks
- extract InviteForm to separate component
- clean up extraneous odds and ends
